chore: reorder constant files and prevent them to be in test coverage
