+++ src/core/constants/countryCodes.ts
@@ -1,4 +1,3 @@
-import { BasicComboBoxData } from '~/components/form'
 import { CountryCode } from '~/generated/graphql'
 
 type countryCodesType = {
@@ -256,12 +255,3 @@ export const CountryCodes: countryCodesType = {
   ZM: 'Zambia',
   ZW: 'Zimbabwe',
 }
-
-export const countryDataForCombobox: BasicComboBoxData[] = (
-  Object.keys(CountryCodes) as Array<keyof typeof CountryCodes>
-).map((countryKey) => {
-  return {
-    value: countryKey,
-    label: CountryCodes[countryKey],
-  }
-})
