+++ src/core/constants/externalUrls.ts
