+++ src/core/constants/globalTypes.ts
